Change use of HOC for a React Hook
